Made payBefore and changed a file name to inventory(too vague) to Sales.txt.
Need to learn how to read and write to files next.
As a lesser priority, I still need to make receipt a constructor method, instead of an instance method. I might just make a differing branch from the master branch and work on it from there, as to not interfere with the working program.